Simplify TEMPO backend sturcture
