+++ oqupy/config.py
@@ -36,13 +36,8 @@ PROGRESS_TYPE = 'bar'
 
 # -- TEMPO --------------------------------------------------------------------
 
-# The default tempo backend for tensor network calculations
-TEMPO_BACKEND = 'tensor-network'
-
-# Dict of all tempo backends and their default configuration
-TEMPO_BACKEND_CONFIG = {
-    'tensor-network': {'backend':'numpy'},
-    }
+# default TEMPO backend configuration
+TEMPO_BACKEND_CONFIG = {}
 
 # maximal dkmax for tempo parameter guessing function
 MAX_DKMAX = 256
@@ -53,13 +48,8 @@ DEFAULT_TOLERANCE = 3.9e-3
 
 # -- PT_TEMPO -----------------------------------------------------------------
 
-# The default process tensor tempo backend for tensor network calculations
-PT_TEMPO_BACKEND = 'tensor-network'
-
-# Dict of all process tensor tempo backends and their default configuration
-PT_TEMPO_BACKEND_CONFIG = {
-    'tensor-network': {},
-    }
+# default PT-TEMPO backend configuration
+PT_TEMPO_BACKEND_CONFIG = {}
 
 # maximal dkmax for process tensor tempo parameter guessing function
 PT_MAX_DKMAX = 256
